Change to int

+++ version_1/Trading-System/src/main/java/TradingSystem/Server/DomainLayer/StoreComponent/Policies/SaleExp/PriceForGetSale.java
@@ -8,9 +8,9 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public class PriceForGetSale extends SimpleExpression {
 
-    Double priceForSale;
+    Integer priceForSale;
 
-    public PriceForGetSale(Double priceForSale) {
+    public PriceForGetSale(Integer priceForSale) {
         this.priceForSale = priceForSale;
     }
