+++ version_1/Trading-System/src/main/java/TradingSystem/Server/DomainLayer/StoreComponent/Policies/Sales/StoreSale.java
@@ -10,16 +10,16 @@ import java.util.concurrent.ConcurrentHashMap;
 public class StoreSale extends SimpleSale {
 
     private Integer storeID;
-    private Double  discountPercentage;
+    private Integer  discountPercentage;
     public TradingSystemImpl tradingSystem=TradingSystemImpl.getInstance();
 
-    public StoreSale(Expression exp,Integer storeID, Double discountPercentage) {
+    public StoreSale(Expression exp,Integer storeID, Integer discountPercentage) {
         super(exp);
         this.storeID = storeID;
         this.discountPercentage = discountPercentage;
     }
 
-    public StoreSale(Integer storeID, Double discountPercentage) {
+    public StoreSale(Integer storeID, Integer discountPercentage) {
         this.storeID = storeID;
         this.discountPercentage = discountPercentage;
     }
